Refactor out runner() function
